Refactor plugins; Highlight debug-assertions in Cargo.toml
