+++ src/default/loading.rs
@@ -1,4 +1,4 @@
-use crate::GameState;
+use crate::default::state::GameState;
 use bevy::prelude::*;
 use bevy_asset_loader::prelude::*;
 use bevy_kira_audio::AudioSource;
