+++ src/default/mod.rs
@@ -1,0 +1,9 @@
+mod actions;
+mod audio;
+mod loading;
+mod menu;
+mod player;
+mod window;
+
+pub mod plugins;
+pub mod state;
