+++ src/default/state.rs
@@ -1,0 +1,12 @@
+// This example game uses States to separate logic
+// See https://bevy-cheatbook.github.io/programming/states.html
+// Or https://github.com/bevyengine/bevy/blob/main/examples/ecs/state.rs
+#[derive(Clone, Eq, PartialEq, Debug, Hash)]
+pub enum GameState {
+    // During the loading State the LoadingPlugin will load our assets
+    Loading,
+    // During this State the actual game logic is executed
+    Playing,
+    // Here the menu is drawn and waiting for player interaction
+    Menu,
+}
